Project building

+++ src/main/scala/com/cherit/models/MovieRatingAccumulator.scala
@@ -1,3 +1,3 @@
 package com.cherit.models
 
-case class MovieRatingAccumulator(movieId: String, ratingCount: Int, ratingSum: Long, ratingUsers: Set[String])
+case class MovieRatingAccumulator(movieId: String, ratingCount: Long, ratingSum: Long, ratingUsers: Set[String])
