+++ src/main/scala/com/cherit/models/MovieRatingAggregated.scala
@@ -1,3 +1,3 @@
 package com.cherit.models
 
-case class MovieRatingAggregated(movieId: String, ratingCount: Int, ratingSum: Int, uniqueRatingCount: Int)
+case class MovieRatingAggregated(movieId: String, ratingCount: Long, ratingSum: Long, uniqueRatingCount: Long)
